Vif: Rename/remove new prefix from files.
