+++ pcsx2/Vif_Dynarec.h
@@ -43,4 +43,4 @@ alignas(16) extern nVifStruct nVif[2];
 alignas(16) extern nVifCall nVifUpk[(2 * 2 * 16) * 4]; // ([USN][Masking][Unpack Type]) [curCycle]
 alignas(16) extern u32      nVifMask[3][4][4];         // [MaskNumber][CycleNumber][Vector]
 
-static constexpr bool newVifDynaRec = 1; // Use code in newVif_Dynarec.inl
+static constexpr bool newVifDynaRec = 1; // Use code in Vif_Dynarec.inl
