+++ pcsx2/arm64/Vif_UnpackNEON.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2002-2024 PCSX2 Dev Team
 // SPDX-License-Identifier: GPL-3.0
 
-#include "newVif_UnpackNEON.h"
+#include "Vif_UnpackNEON.h"
 #include "common/Perf.h"
 
 namespace a64 = vixl::aarch64;
