+++ pcsx2/x86/Vif_UnpackSSE.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2002-2024 PCSX2 Dev Team
 // SPDX-License-Identifier: GPL-3.0+
 
-#include "newVif_UnpackSSE.h"
+#include "Vif_UnpackSSE.h"
 #include "common/Perf.h"
 #include "fmt/core.h"
